requirements.txt
alembic==1.6.2
aniso8601==7.0.0
attrs==21.2.0
bleach==3.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-plugins==1.1.1
cligj==0.7.1
coloredlogs==14.0
croniter==1.0.13
dagit==0.11.9
dagster==0.11.9
dagster-graphql==0.11.9
defusedxml==0.7.1
docstring-parser==0.7.1
entrypoints==0.3
Fiona==1.8.19
Flask==1.1.4
Flask-Cors==3.0.10
Flask-GraphQL==2.0.1
Flask-Sockets==0.2.1
future==0.18.2
geopandas==0.9.0
gevent==21.1.2
gevent-websocket==0.10.1
gql==2.0.0
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
graphql-ws==0.3.1
greenlet==1.1.0
grpcio==1.37.1
grpcio-health-checking==1.37.1
humanfriendly==9.1
idna==2.10
ipython-genutils==0.2.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter-core==4.7.1
Mako==1.1.4
MarkupSafe==2.0.0
mistune==0.8.4
munch==2.5.0
nbconvert==5.6.1
nbformat==5.1.3
numpy==1.20.3
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
pendulum==2.1.2
Pillow==8.2.0
promise==2.3
protobuf==3.17.0
Pygments==2.9.0
pyparsing==2.4.7
pyproj==3.0.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
pytzdata==2020.1
PyYAML==5.4.1
requests==2.25.1
Rx==1.6.1
Shapely==1.7.1
six==1.16.0
SQLAlchemy==1.4.15
tabulate==0.8.9
testpath==0.5.0
toposort==1.6
tqdm==4.60.0
traitlets==5.0.5
urllib3==1.26.4
watchdog==2.1.1
webencodings==0.5.1
Werkzeug==1.0.1
zope.event==4.5.0
zope.interface==5.4.0
